CLJS-2592: :npm-deps using ES6 modules with .mjs extensions are not detected correctly
This adds a new :package-json-resolution option that can be used to
control which package.json entry fields are considered during JS module
resolution. As of this commit, the option accepts three values:

    :webpack (for ["browser" "module" "main"])
    :nodejs (for ["main"])
    a custom vector of entries (e.g. ["main" "module"])

If the :package-json-resolution option is not set, we fall back to the
two default modes based on the build target:

  No target => :webpack
  Target :nodejs => :nodejs

The resulting behavior:

    {:target :nodejs}  => ["main"]
    {} => ["browser", "module", "main"]

    {:target <any>
     :package-json-resolution :webpack} => ["browser", "module", "main"]

    {:target <any>
     :package-json-resolution :nodejs} => ["main"]

    {:target <any>
     :package-json-resolution ["foo", "bar"]} => ["foo", "bar"]

Using the new :nodejs mode, this allows to import packages like graphql
and iterall via :npm-deps, which currently point their "module" entries
to unsupported .mjs files.