CLJS-2641: cljs.spec.alpha/fdef with s/* is broken
The problem was while `instrument` redefs the var, the original fn
internals still self references through the namespace, thus
triggering an infinite loop by calling into the instrumented version
again.

This change relies on the fact that `apply` will always bind the
original fn via `Function.prototype.call` when invoking. In the dispatch
helpers we now check if `this` is bound to a Function - if so we invoke
methods on that value instead as it will be the *original* fn.