(WIP)
Out is now directed to the correct place but there are two outstanding
issues. The first is that the simple `pipe` no longer works as the
value is now JSON. As we depend on Transit already we could use the
streaming Jackson JSON parser to address this.

The second issue is that since we pipe, the observed order in REPLs
is non-deterministic.