CLJS-2650: Fix JAR compilation of cljs.loader
The cljs.loader file needs to be compiled after all other files to
populate the `module-infos` and `module-uris` vars. The
`requires-compilation?` function returned false if the file to compile
was cljs.loader and didn't contain a :cache-key. This prevented
`compile-from-jar` from moving the cljs.loader from the JAR file to
the disk and caused a java.io.FileNotFoundException later on.

Since cljs.loader always get compiled in a later pass, we decided to
remove this special logic.

Since this behaviour only happens when compiling from a JAR, I added a
test that uses the cljs.jar to compile the example from the Code
Splitting guide.