Recompile cljs.loader in REPL
The cljs.loader namespace has to be compiled last, because all inputs
need to be known to populate the `MODULE_INFOS` and `MODULE_URIS`
vars.

Loading a file or evaluating a form in the REPL could trigger another
compilation of the cljs.loader namespace, without the information for
the `MODULE_INFOS` and `MODULE_URIS` available. This compilation
erased the previously populated `MODULE_INFOS` and `MODULE_URIS` vars
and corrupted the module loader.

This patch modifies the loader compilation pass to swap and merge the
`MODULE_INFOS` and `MODULE_URIS` into the compiler state, to make this
information available to compilation stages happening later.

The patch also changes the REPL to recompile and load the cljs.loader
namespace when a file gets loaded, or a namespace form gets
evaluated. A new module graph will be calculated and merged into the
previously calculated one. Recompilation of the cljs.loader namespace
will only happen if the file or form require the cljs.loader
namespace.

Since the :aot-cache is populated at a point where the information for
the `MODULE_INFOS` and `MODULE_URIS` is not yet available, this patch
also makes sure the cljs.loader namespace is never cached.