CLJS-2746 Missing provides for JS modules
Prior to this change, JS modules are converted my writing a single
goog.provide statement that corresponds to the file path and name of the
module. The .getProvides method (of another module) may use a shorter
name and this means that the ClojureScript compiler can never find the
module by this shorter name.

This change emits a goog.provide statement for each name specified in
the provides list of the IJavaScriptFile.