CLJS-2775: cljs.main: Node modules not installed if -re node
If you look at the (non-cljs.main) REPL startup sequence in
cljs.repl/repl*, maybe-install-npm-deps is called prior to run-inits,
but in cljs.cli, there is a call to run-inits without first ensuring
that maybe-install-npm-deps has been called. This is simple to fix by
explicitly making such a call before run-inits is called.