CLJS-2780: Async tests prematurely terminate in Node
While the current implementation of -tear-down in the Node REPL allows
any pending (Node termination delaying) processing to finish, it
doesn't allow for any further communication between the REPL and the
Node process: It immediately destroys the Node-side socket while also
discarding any REPL-side communication structures.

This patch moves the code that discards REPL-side communication
structures to occur after awaiting Node process termination so that any
messages that are sent can be displayed. It also replaces the Node-side
destroy call to an unref: This allows any lingering communication to
occur, without causing the socket to keep the Node process alive.