CLJS-2798: ChunkCons -next doesn't handle nil more
This revison aligns more with the Clojure implementation.

In the Clojure implemenentation, if _more is nil, then it would end up
doing a seq on an empty list, which would produce nil. This revision
simply pre-checks if more is nil, and if not, returns nil in that case.

The change also eliminates an unnecessary check of the return value of
-seq (checking for nil and returning nil), and just returns the value
-seq produces.