CLJS-2693: Have Range implement IChunkedSeq
This set of changes mirrors those in Clojure surrounding LongRange and
LongChunk. In particular, range is modified to follow the 3-arg
LongRange/create.

The previous next implementation is preserved. (The Clojure
implemenation forces a chunk for each next call, which, in
ClojureScript, slows down algorithms like doall which walk the seq.)

RangeChunk does not implement IReduce because, unlike in Clojure,
ClojureScript doesn't have an internal-reduce which reduces chunked-seq
chunks. (Additionally, reduce for ranges is handled directly by Range.)