CLJS-2901: Return type inference for multi-arity & variadic fns
For the anonymous fn case simply handle the tagging of :methods in
cljs.analyzer/infer-type. This is enough for local fn invoke to be
be inferred correctly.

The top-level case is a bit trickier due to the fact that defn
is a ClojureScript macro that desugars before the compiler to ensure
that Closure cross module code motion can work. In this case add a
special case to cljs.analyzer/parse 'set!. If we have a top level
fn (:top-fn), use its metadata to recover the original fn information
pre-desugaring. :tag information comes from analyzing the body.

cljs.analyzer/infer-invoke has been changed to prefer inference over
:ret-tag.

In all cases we make sure enough `:methods` information is available
for infer-invoke to work.

Tweak cljs.analyzer/analyze-form-seq so that it's a bit more useful
under testing.

Add basic tests cases for the above as well as bringing back a large
number of inference tests that were commented out.