CLJS-3068: Compiler error with if and emit-var
The test-induced tags implementation is set up to inject induced
tags on locals, but the truthy-induced code inadvertently
considered a broader scope (any simple symbol, which could be
a symbol which resolves to a var).

This results in a defect where a new entry is added to :locals for a
var, and the fix is to properly limit the scope for truthy-induced.