CLJS-3144: NPM Modules should have all their vars marked to avoid .call invokes
add new var information to vars resolved from Node or global foreign modules -
mark them as :foreign true.

In the compiler if a f is :foreign, then do not emit the .call convention.
By doing this we avoid issues when the library itself is a singleton
instance as .call convention breaks `this`.