CLJS-3219: Problem with asserts namespace inside goog.math.Long
move more shareable logic into cljs.repl.bootstrap - do the goog.isProvided_
monkeypatch there - go ahead and keep a reference to the original just in case we
need it in the future.

tweak goog.require monkey patch to check if we're in the goog.module loader
and in that case return the module. This is only for goog.module - nothing
we would ever generate in ClojureScript