CLJS-712 & CLJS-2957: resolve-var for dot still wrong, externs inference regression
The AST alignment changes reveal that our handling of the dot form was still
incorrect. When resolve-var returns a dotted :var like foo.bar/baz.woz it gets
desugared. In this case of GCL import this would lead to malformed :host-field
AST nodes.

Now in resolve var if we have a dotted symbol first we check that the prefix
is *not* something that can be resolved. If can't be resolved then we don't
do anything other than convert the form from foo.bar.baz to foo.bar/baz.

Add tests cases for type inference, externs inference, and code gen.