CLJS-3244: Warnings for clojure.browser.net with :static-fns true
In :invoke emission of the compiler we were resolving the tag to
check to see if the type implements a protocol. But this could
easily be a type we can't possibly know about - i.e. Google
Closure Library ctor, foreign library ctor, etc.

Explicit type-hinting is not idiomatic over predicate driven
optimization, so just suppress any warnings.