CLJS-1147: Reconnect logic for browser REPL
Add an error handler on the xhr-connection in the evaluator process.
If this fails, set state as disconnected, notify parent of disconnect
and retry after a 1 second delay.

Remove the timeout from try-handshake now that this logic is in place.

Now browser REPL works in all ordering scenarios. A browser can be closed
and reopened, a REPL can be stopped and a new one started to reconnect
to an open browser w/o refresh.