CLJS-3276 Support macros that expand to require statements
Support this pattern:

   (ns foo.bar
     (:require-macros [foo.baz :refer [macro-that-expands-to-require]]))
   (macro-that-expands-to-require)

To do so we adapt ana/parse-ns to look for consecutive :ns/:ns* ops (ns and
require forms), until a non :ns/:ns* form is encountered. The information of
consecutive forms are merged so the ns-info that parse-ns returns is complete.
This was already happening for :ns* ops (require forms), but not for :ns forms.

This means that parse-ns analyzes/parses at least one form beyond the last
ns/require form, but this can be problematic since this modifies the compiler
env. For instance if the form after ns is a (def ^const ...) then analyzing this
twice will cause an error. Hence the check to only analyze macro forms or
ns/require forms. This should also help in general to avoid unnecessary work.

To make sure the invocation is seen as a macro, and is able to be expanded, we
need to keep track of namespace information in the env across multiple :ns/:ns*
ops. When encountering a macro invocation we load the macro ns on the fly by
invoking the ns-side-effects analyzer pass directly. After this we can call
analyze on the form.