CLJS-3346: :as-alias (#119)
Implement `:as-alias` support as simple structural pass on the ns form EDN. All `:as-alias` are collected, and the ns libspecs are returned with `:as-alias` elided. If a lib spec has *only* `:as-alias` it is *removed* entirely. This is possible since `:as-alias` does not load.

The structural pass approach means further alterations to the analyzer or compiler are not needed. This aligns w/ the purpose of `:as-alias`, as it is primarily about feeding the reader.

One interesting thing to note is that ClojureScript has two difference namespaces - one for defs and one for macros and they may in fact collide. Since `:as-alias` is only about the reader - collisions are not allowed across defs and macros. That is, if a `:require-macros` libspec declares an `:as-alias` it cannot be reused by a `:require` libspec.

* change analyzer `ns` and `ns*` to support `:as-alias`
* change ns merging to deep merge `:as-aliases` 
* change `form-seq` reader helper to include `:as-aliases` from the current namespace
* change REPL reader to include `:as-aliases` from the current namespace