CLJS-3347: Create clojure.math namespace
This introduces the cljs.math namespace and clojure.core/abs for
parity with Clojure 1.11.0-alpha4.
Also introduces clojure.math-tests, which is an equivalent sets of
ClojureScript tests to the associated tests in Clojure.
Also include the Clojure test clojure.gen-math-test to compare
all reimplemented functions between the JVM versions and the
equivalent function calls executed on a ClojureScript prepl.