CLJS-3373: Externs Inference issue with vars invoked from foreign libs (#173)
Fix invoke inference to handle invokes on vars from foreign libs. If
we have a :js-var we cannot know whether it is a function or property.
Note this is different from the `:js-fn-var` case where a provided
extern did disambiguate.

In the case of `:js-var` we throw away the leading prefix since the
types simply cannot be known.

Add a test case based on the one provided by Timothy Pratley.