Experimental Lite Mode (#265)
new code-size compiler flags: :lite-mode, :elide-to-string

:lite-mode emits updated variants of the original copy-on-write data
structures from 2011. Notably, in `:lite-mode` chunked-seqs are not
supported as they pull in a significant amount of code, which defeats
the purpose of the new flags. Also while the transient interfaces are
implemented on the copy-on-write data structures, they do not actually
provide a fast path. This is not by design, but simply a practical
scoping of effort as the work done here is already considerable.

The other serious code-size issue are `.toString` impls. There are
many interesting smaller programs that simply do not need the
convenient `.toString` invoking the ClojureScript recursive printing
machinery. By combining `:lite-mode true` with `:elide-to-string true`
the emitted code is cut in half or more for simpler programs.

For a rough idea of the space-savings, a hash-map literal is ~6K after
advanced compilation with the new flags and brotli
compression. ClojureScript 1.12.42 is ~17K brotli for the same program.

The following expression

(.log js/console
      (->> (map inc (range 10))
           (filter even?)
           (partition 2)
           (drop 1)
           (mapcat identity)
           into-array))

Is ~6K after advanced compilation + brotli compression. ClojureScript
is 1.12.42 is ~19K.

For users that want the full expressiveness of the Clojure APIs and
the interactivity of the REPL for smaller and/or simpler projects, but
don't need every last bit of performance, these two flags should lead
to compact artifacts.